feature: pydantic refactor
